// submodules/common-dto/api-client/main/models/PayoutStateEnum.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * The status of the payment
 */
export enum PayoutStateEnum {
    CREATED = 'created',
    ACCEPTED = 'accepted',
    REJECTED = 'rejected',
    RELEASED = 'released',
    RELEASE_FAILED = 'release_failed',
    CANCELLED = 'cancelled',
}